add command to button
relocate query to command
